it works now
