+++ app/routes/api/posts/$postId/comments/index.ts
@@ -1,10 +1,9 @@
 import type { ActionFunction } from "@remix-run/node";
 import { json } from "@remix-run/node";
-import { comment } from "postcss";
 import invariant from "tiny-invariant";
 import { createComment } from "~/models/post-comments.server";
 
-import { requireUserId } from "~/services/session.server";
+import { requireUserId } from "~/services/auth.server";
 
 export const action: ActionFunction = async ({ request, params }) => {
   if (request.method !== "POST") {
