+++ app/routes/api/posts/$postId/reactions.ts
@@ -2,7 +2,7 @@ import type { ActionFunction } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import invariant from "tiny-invariant";
 
-import { requireUserId } from "~/services/session.server";
+import { requireUserId } from "~/services/auth.server";
 import { isEmoji } from "~/lib/emoji";
 import { togglePostReaction } from "~/models/post-reactions.server";
