+++ app/routes/api/posts/$postId/subscription.ts
@@ -1,12 +1,12 @@
 import type { ActionFunction } from "@remix-run/node";
 import { json } from "@remix-run/node";
 import invariant from "tiny-invariant";
+
 import {
   createSubscription,
   deleteSubscription,
 } from "~/models/post-subscription.server";
-
-import { requireUserId } from "~/services/session.server";
+import { requireUserId } from "~/services/auth.server";
 
 export const action: ActionFunction = async ({ request, params }) => {
   if (request.method !== "DELETE" && request.method !== "POST") {
