+++ app/routes/c/$categorySlug.tsx
@@ -3,7 +3,7 @@ import { json } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import invariant from "tiny-invariant";
 
-import { requireUserId } from "~/services/session.server";
+import { requireUserId } from "~/services/auth.server";
 import { getPostList } from "~/models/post.server";
 import { getCategory } from "~/models/category.server";
 import Post from "~/components/post";
