+++ app/routes/feeds/$feedId[.]xml.tsx
@@ -36,7 +36,7 @@ export const loader: LoaderFunction = async ({ request, params }) => {
               <content:encoded><![CDATA[${escapeCdata(
                 marked.parse(post.content as string, { breaks: true })
               )}]]></content:encoded>
-              <author>${escapeHtml(post.author.name)}></author>
+              <author>${escapeHtml(post.author.name)}</author>
               <pubDate>${post.publishedAt.toUTCString()}</pubDate>
               <link>${buildUrl(request, getPostLink(post))}</link>
               <guid>${post.id}</guid>
